feat(cmd/tools): add 'protoc' command and
and upgrade protoc & add linux arm support.

* refactor data dirs for protoc and nodetime pkgs.
