+++ .goreleaser.yml
@@ -14,3 +14,4 @@ brews:
       name: "homebrew-tap"
     install: |
       bin.install "starport"
+
