+++ starport/pkg/cosmosgen/generate_javascript.go
@@ -12,9 +12,9 @@ import (
 	"github.com/tendermint/starport/starport/pkg/cosmosanalysis/module"
 	"github.com/tendermint/starport/starport/pkg/giturl"
 	"github.com/tendermint/starport/starport/pkg/gomodulepath"
-	"github.com/tendermint/starport/starport/pkg/nodetime/sta"
-	tsproto "github.com/tendermint/starport/starport/pkg/nodetime/ts-proto"
-	"github.com/tendermint/starport/starport/pkg/nodetime/tsc"
+	"github.com/tendermint/starport/starport/pkg/nodetime/programs/sta"
+	tsproto "github.com/tendermint/starport/starport/pkg/nodetime/programs/ts-proto"
+	"github.com/tendermint/starport/starport/pkg/nodetime/programs/tsc"
 	"github.com/tendermint/starport/starport/pkg/protoc"
 	"github.com/tendermint/starport/starport/pkg/xstrings"
 	"golang.org/x/sync/errgroup"
