+++ starport/pkg/cosmosgen/generate_openapi.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/iancoleman/strcase"
 	"github.com/tendermint/starport/starport/pkg/cosmosanalysis/module"
-	swaggercombine "github.com/tendermint/starport/starport/pkg/nodetime/swagger-combine"
+	swaggercombine "github.com/tendermint/starport/starport/pkg/nodetime/programs/swagger-combine"
 	"github.com/tendermint/starport/starport/pkg/protoc"
 )
