--- starport/pkg/nodetime/binary_darwin_amd64.go
@@ -1,8 +1,0 @@
-// +build darwin amd64
-
-package nodetime
-
-import _ "embed" // embed is required for binary embedding.
-
-//go:embed nodetime-darwin-amd64.tar.gz
-var binaryCompressed []byte
