--- starport/pkg/nodetime/binary_linux_amd64.go
@@ -1,8 +1,0 @@
-// +build linux amd64
-
-package nodetime
-
-import _ "embed" // embed is required for binary embedding.
-
-//go:embed nodetime-linux-amd64.tar.gz
-var binaryCompressed []byte
