+++ starport/pkg/nodetime/data/data.go
@@ -1,0 +1,6 @@
+package data
+
+// Binary returns the compressed platform spesific nodetime binary.
+func Binary() []byte {
+	return binaryCompressed
+}
