+++ starport/pkg/nodetime/data/data_linux_amd64.go
@@ -1,0 +1,6 @@
+package data
+
+import _ "embed" // embed is required for binary embedding.
+
+//go:embed nodetime-linux-amd64.tar.gz
+var binaryCompressed []byte
