--- starport/pkg/protoc/binary_darwin.go
@@ -1,8 +1,0 @@
-// +build darwin
-
-package protoc
-
-import _ "embed" // embed is required for binary embedding.
-
-//go:embed data/protoc-darwin-amd64
-var binary []byte
