+++ starport/pkg/protoc/data/data_darwin_amd64.go
@@ -1,0 +1,6 @@
+package data
+
+import _ "embed" // embed is required for binary embedding.
+
+//go:embed protoc-darwin-amd64
+var binary []byte
