+++ starport/pkg/protoc/data/data_linux_amd64.go
@@ -1,0 +1,6 @@
+package data
+
+import _ "embed" // embed is required for binary embedding.
+
+//go:embed protoc-linux-amd64
+var binary []byte
